change stock

+++ src/server/controller/stock_storehouse.js
@@ -1,13 +1,13 @@
-const Base = require('./base.js');
+const stockBase = require('./stock_base.js');
 /**
  * @class
  * @apiDefine stock_storehouse 仓库管理管理
  */
-module.exports = class extends Base {
+module.exports = class extends stockBase {
 
     async listAction() {
         let { page, limit, param } = this.get();
-        let wsql = {};
+        let wsql = {group_id : this.groupId};
         if (param) wsql = this.parseSearch(param, wsql);
         let list = await this.model('stock_storehouse').where(wsql).page(page, limit).order('id desc').select();
         let count = await this.model('stock_storehouse').where(wsql).count();
@@ -16,7 +16,10 @@ module.exports = class extends Base {
 
     async addAction() {
         let post = this.post();
+		post.group_id = this.groupId;
+		post.user_id = this.adminId;
         let id = await this.model('stock_storehouse').add(post);
+		await this.upStorehouseCache();
         return this.success(id);
     }
 
@@ -25,6 +28,7 @@ module.exports = class extends Base {
         let has = await this.model('stock_storehouse').where({ id: post.id }).find();
         if (think.isEmpty(has)) return this.fail('编辑的数据不存在');
         await this.model('stock_storehouse').update(post);
+		await this.upStorehouseCache();
         return this.success()
     }
 
@@ -39,7 +43,11 @@ module.exports = class extends Base {
         let id = this.post('id');
         if (!await this.hasData('stock_storehouse', { id }))
             return this.fail('数据不存在')
+		let hasBar = await this.model('stock_bar').where({area_id : id}).find()
+		if(!think.isEmpty(hasBar))
+			return this.fail('仓库下存在货架数据')
         await this.model('stock_storehouse').where({ id }).delete()
+		await this.upStorehouseCache();
         return this.success()
     }
 }
